Fix remaining breaks post nonzero range cfg change

+++ examples/reshare/src/dkg/actor/dealer.rs
@@ -97,7 +97,7 @@ where
             ctx.with_label("storage"),
             storage_partition,
             round_info.round(),
-            round_info.max_read_size().get(),
+            round_info.max_read_size().get() as usize,
         )
         .await;
         let mut ctx = ContextCell::new(ctx);
